Move logging tests to more appropriate file.
These functions were moved, so the tests should move too. They have also been updated to match the changes in the functions.
