+++ tests/core.py
@@ -25,8 +25,6 @@ import tempfile
 from datetime import datetime, time, timedelta
 from email.mime.multipart import MIMEMultipart
 from email.mime.application import MIMEApplication
-from io import StringIO
-from tempfile import NamedTemporaryFile
 from time import sleep
 import warnings
 
@@ -56,9 +54,7 @@ from airflow.www import app as application
 from airflow.settings import Session
 from airflow.utils.state import State
 from airflow.utils.dates import round_time
-from airflow.utils.logging import LoggingMixin,\
-                                  setup_stream_logging,\
-                                  setup_file_logging
+from airflow.utils.logging import LoggingMixin
 from lxml import html
 from airflow.exceptions import AirflowException
 from airflow.configuration import AirflowConfigException
@@ -905,30 +901,6 @@ class CoreTest(unittest.TestCase):
         session.query(models.DagStat).delete()
         session.commit()
         session.close()
-
-
-class CliLoggingTests(unittest.TestCase):
-    def tearDown(self):
-        if self.handler:
-            logging.getLogger().removeHandler(self.handler)
-
-    def test_setup_stream_logging(self):
-        # Make sure our handler is getting messages.
-        self.handler = setup_stream_logging()
-        stream = StringIO()
-        self.handler.stream = stream  # Override stderr default stream.
-        logger = logging.getLogger()
-        logger.info("test message")
-        self.assertIn("test message", stream.getvalue())
-
-    def test_setup_file_logging(self):
-        with NamedTemporaryFile('w+t') as tempfile:
-            self.handler = setup_file_logging(tempfile.name)
-            logger = logging.getLogger()
-            logger.info("test message")
-            tempfile.seek(0)
-            log_message = tempfile.read()
-            self.assertIn("test message", log_message)
 
 
 class CliTests(unittest.TestCase):
